Support recent changes to vdso functions
In newer versions of glibc, we encounter a CMP instruction with a
imm16/32 operand (opcode 0x81) when detouring vdso functions. CMP with a
imm8 operand (opcode 0x83) is already supported and all the logic is the
same for this.

This also adds more useful output when an opcode is not supported.